Adds two-few-points failure to Simplify.

+++ parse.bash
@@ -1,5 +1,6 @@
 #!/bin/bash
 set -x
+set -eu
 rm -fr ./wgs84 ./xy ./png
 awk -f ./parse_tasks.awk taskdata.xml
 for i in *.wgs84; do
@@ -14,8 +15,11 @@ for i in *.wgs84; do
   gnuplot -e "outfile='${base}_defl.png'; titletxt='$title'" deflect.plt
   gnuplot -e "outfile='${base}_hist.png'; titletxt='$title'" histo.plt
 done
-rm -f ~/storage/downloads/ekman/*.png
-cp -pv *.png ~/storage/downloads/ekman
+if test -d ~/storage/downloads/ekman
+then
+  rm -f ~/storage/downloads/ekman/*.png
+  cp -pv *.png ~/storage/downloads/ekman
+fi
 rm points.xy out.xy corners.xy deflect.txt
 mkdir -p wgs84
 mv *.wgs84 wgs84
